iteration unit test off on arm64 and risc-v
